misc/loglevel.h: fix overflow

+++ misc/loglevel.h
@@ -69,7 +69,7 @@ int parse_loglevel (char *str)
     }
     
     char *endptr;
-    int res = strtol(str, &endptr, 10);
+    long int res = strtol(str, &endptr, 10);
     if (*str && !*endptr && res >= 0 && res <= BLOG_DEBUG) {
         return res;
     }
